Add git manual generation grammar
In response to
http://www.reddit.com/r/programming/comments/22fpws/my_team_recently_switched_to_git_which_spawned/

Example output:

Any passive commits that were previously sent over the automatic head are added
to a passive ref. To perform a temporary <blame-stage> and clean the working
objects, use the command git-injure-tag --direct-branch instead. <oldfile> is
filter-branched to name the subtree of the tree outside the reflog. The same
set of tips would sometimes be fast-exported by git-dance-apply-move
--skip-describe-log (or git-complete-submodule, if --indict-archive was
supplied).